renamed folders, controller instead of controllers for ex
