+++ lib/client/controller/firstrun.rb
