+++ lib/client/controller/status.rb
