+++ lib/client/controller/update.rb
