+++ lib/client/helper/formatting.rb
