+++ lib/client/helper/results.rb
