+++ lib/client/helper/time.rb
