+++ lib/client/model/endpoint.rb
